Update Landsat 8 Collection 2 Level 2 asset keys
The Landsat 8 Collection 2 Level 2 created item asset keys currently
has the file extension in the key name. With the current code, there
was inconsistent capatilization of the extension. On consideration,
the inclusion of file extensions is consistent with the USGS 0.7
Items but is inconsistent with other STAC items produces. The assets
themselves describe the media type and the HREFs have the file
extension, and having the file extension in asset key is duplicate
information and may be confusing to users. This
commit switches to dropping file extensions in key names. The only
exception is the MTL file, which has multiple file types - these
assets retain their extension to differentiate between the various
assets.